refactor: Move BaseMiddleware to _internal module.
