+++ faststream/_internal/middlewares.py
@@ -1,12 +1,9 @@
 from collections.abc import Awaitable
 from typing import TYPE_CHECKING, Any, Callable, Generic, Optional
 
-from typing_extensions import (
-    Self,
-    TypeVar as TypeVar313,
-)
+from typing_extensions import Self
 
-from faststream.response import PublishCommand
+from faststream._internal.types import PublishCommandType
 
 if TYPE_CHECKING:
     from types import TracebackType
@@ -14,13 +11,6 @@ if TYPE_CHECKING:
     from faststream._internal.basic_types import AsyncFuncAny
     from faststream._internal.context.repository import ContextRepo
     from faststream.message import StreamMessage
-
-
-PublishCommandType = TypeVar313(
-    "PublishCommandType",
-    bound=PublishCommand,
-    default=PublishCommand,
-)
 
 
 class BaseMiddleware(Generic[PublishCommandType]):
