+++ faststream/middlewares/__init__.py
@@ -1,6 +1,6 @@
+from faststream._internal.middlewares import BaseMiddleware
 from faststream.middlewares.acknowledgement.conf import AckPolicy
 from faststream.middlewares.acknowledgement.middleware import AcknowledgementMiddleware
-from faststream.middlewares.base import BaseMiddleware
 from faststream.middlewares.exception import ExceptionMiddleware
 
 __all__ = (
