+++ faststream/response/__init__.py
@@ -1,8 +1,9 @@
 from .publish_type import PublishType
-from .response import PublishCommand, Response
+from .response import BatchPublishCommand, PublishCommand, Response
 from .utils import ensure_response
 
 __all__ = (
+    "BatchPublishCommand",
     "PublishCommand",
     "PublishType",
     "Response",
